requirements/production.txt
# Include base requirements
-r base.txt

# Production server
gunicorn==23.0.0

# Redis cache backend (celery and redis already in base.txt)
django-redis==6.0.0

# Security & monitoring
django-environ==0.12.0
sentry-sdk==2.49.0

# Database URL parsing (Railway/Heroku support)
dj-database-url==3.1.0
